- Replace media type emoji with labels. - Don't speak out metrics overlay. - Increase the height of the slider. - Speak out the slider progress as a percentage. - Add state label to the player control view.
