+++ pillarbox-demo-shared/src/main/res/values/strings.xml
@@ -11,6 +11,10 @@
     <string name="search_placeholder">Search for content</string>
     <string name="no_results">No results</string>
     <string name="empty_search_query">Enter something to search</string>
+    <string name="audio_content">Audio</string>
+    <string name="video_content">Video</string>
+    <string name="controls_visible">Controls visible</string>
+    <string name="controls_hidden">Controls hidden</string>
     <string name="duration"><xliff:g example="15" id="duration">%1$d</xliff:g> min</string>
     <string name="settings">Settings</string>
     <string name="audio_track">Audio tracks</string>
